+++ src/assets/scripts/client/aircraft/FlightManagementSystem/Fms.js
@@ -1,4 +1,5 @@
 import _find from 'lodash/find';
+import _findLast from 'lodash/findLast';
 import _includes from 'lodash/includes';
 import _isEmpty from 'lodash/isEmpty';
 import _isNil from 'lodash/isNil';
@@ -13,7 +14,7 @@ import {
 } from '../../constants/aircraftConstants';
 import { INVALID_NUMBER } from '../../constants/globalConstants';
 import { PROCEDURE_OR_AIRWAY_SEGMENT_DIVIDER } from '../../constants/routeConstants';
-
+import { radio_runway } from '../../utilities/radioUtilities';
 
 /**
  * Provides methods to create, update or replace a flightPlan and the legs
@@ -791,7 +792,7 @@ export default class Fms {
      * @return {array<boolean, string>}
      */
     replaceArrivalProcedure(routeString) {
-        const routeStringElements = routeString.split(PROCEDURE_OR_AIRWAY_SEGMENT_DIVIDER);
+        const routeStringElements = routeString.toUpperCase().split(PROCEDURE_OR_AIRWAY_SEGMENT_DIVIDER);
 
         if (routeStringElements.length !== 3) {
             return [false, 'arrival procedure format not understood'];
@@ -822,28 +823,68 @@ export default class Fms {
      * @param routeString {string}
      * @return {boolean}
      */
-    replaceDepartureProcedure(routeString) {
-        const routeStringElements = routeString.split(PROCEDURE_OR_AIRWAY_SEGMENT_DIVIDER);
+    replaceDepartureProcedure(routeString, airportIcao) {
+        const routeStringElements = routeString.toUpperCase().split(PROCEDURE_OR_AIRWAY_SEGMENT_DIVIDER);
 
-        if (routeStringElements.length !== 3) {
+        if (routeStringElements.length > 3) {
             return [false, 'departure procedure format not understood'];
         }
 
-        const procedureId = routeStringElements[1];
+        let procedureId = routeStringElements[0];
 
-        if (!NavigationLibrary.hasProcedure(procedureId)) {
+        if (routeStringElements.length === 3) {
+            procedureId = routeStringElements[1];
+        }
+
+        const sidModel = NavigationLibrary.getProcedure(procedureId);
+
+        if (_isNil(sidModel)) {
             return [false, `unknown procedure "${procedureId}"`];
         }
 
-        const wasSuccessful = this._routeModel.replaceDepartureProcedure(routeString);
+        if (routeStringElements.length === 1) { // RouteString looks like PROC
+            const exitPoint = _findLast(this.waypoints, (waypointModel) => sidModel.hasExit(waypointModel.name));
 
-        if (wasSuccessful) {
-            this._updateDepartureRunwayFromRoute();
+            if (!exitPoint) {
+                return [false, `the ${procedureId.toUpperCase()} departure doesn't have an exit along our route`];
+            }
 
-            return [true, ''];
+            routeStringElements.push(exitPoint.name);
         }
 
-        return [false, `route of "${routeString}" is not valid`];
+        if (routeStringElements.length === 2) { // RouteString looks like PROC.EXIT
+            const runwayModel = this.departureRunwayModel;
+
+            if (_isNil(runwayModel)) {
+                return [false, 'unsure if we can accept that procedure; we don\'t have a runway assignment'];
+            }
+
+            let entryName = sidModel.getUniqueEntryPoint();
+
+            if (_isEmpty(entryName)) {
+                entryName = airportIcao.toUpperCase() + runwayModel.name;
+
+                if (!sidModel.hasEntry(entryName)) {
+                    const readback = {
+                        log: `the ${procedureId.toUpperCase()} departure is not valid for runway ${runwayModel.name}`,
+                        say: `the ${procedureId.toUpperCase()} departure is not valid for runway ${radio_runway(runwayModel.name)}`
+                    };
+
+                    return [false, readback];
+                }
+            }
+
+            routeStringElements.unshift(entryName);
+        }
+
+        const nextRouteString = routeStringElements.join(PROCEDURE_OR_AIRWAY_SEGMENT_DIVIDER);
+        const readback = this._routeModel.replaceDepartureProcedure(nextRouteString);
+
+        if (readback[0]) {
+            this._updateDepartureRunwayFromRoute();
+        }
+
+        return readback;
     }
 
     // TODO: we may need to update the runway in this method