+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - [#155](https://github.com/openscope/openscope/issues/155) - Accept entrail property in spawnPatternModel
 - [#1045](https://github.com/openscope/openscope/issues/1045) - Arrivals ignoring altitude restrictions
 - [#1084](https://github.com/openscope/openscope/issues/1084) - Add messages for too-low glideslope intercepts
+- [#955](https://github.com/openscope/openscope/issues/955) - Fix SID Command doesn't work